requirements.txt
aiogram==3.20.0.post0
psycopg2==2.9.10
psycopg2_binary==2.9.10
python-dotenv==1.1.0
